drivers: aw8624_haptic: Fix void-pointer-to-int-cast warning
../drivers/input/misc/aw8624_haptic/aw8624.c:2271:8: warning: cast to smaller integer type 'int' from 'u8 *' (aka 'unsigned char *') [-Wpointer-to-int-cast]
                     (int)effect->pattern);
                     ^~~~~~~~~~~~~~~~~~~~

+++ drivers/input/misc/aw8624_haptic/aw8624.c
@@ -2717,7 +2717,7 @@ static int aw8624_parse_dt(struct device *dev, struct aw8624 *aw8624,
 		pr_debug
 		    ("%s: %d  effect->pattern_length=%d  effect->pattern=%d \n",
 		     __func__, __LINE__, effect->pattern_length,
-		     (int)effect->pattern);
+		     (uintptr_t)effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
